Move functions form VisualResourceNodeInfo into .cpp from .h Add more logs for debugging
